split views in folders
