+++ .pylintrc
@@ -2,7 +2,7 @@
 init-hook='import sys, os; sys.path.append(os.getcwd())'
 
 [MESSAGES CONTROL]
-disable = C0114, C0116, R0903, C0115
+disable = C0114, C0116, R0903, C0115, E1101
 
 # List of comma-separated files or directories to ignore.
 ignore=migrations
