+++ eventplannerdb/views/auth/auth_views.py
@@ -1,0 +1,41 @@
+from django.shortcuts import render, redirect
+from django.views import View
+from django.contrib.auth import login, logout
+from django.contrib.auth.forms import UserCreationForm, AuthenticationForm
+
+class LoginView(View):
+    def get(self, request):
+        if request.user.is_authenticated:
+            # Redirect to event_list page if the user is already logged in
+            return redirect('event_list')
+
+        form = AuthenticationForm()
+        return render(request, "login.html", {"form": form})
+
+    def post(self, request):
+        form = AuthenticationForm(data=request.POST)
+        if form.is_valid():
+            user = form.get_user()
+            login(request, user)
+            return redirect("event_list")
+        return render(request, "login.html", {"form": form})
+
+
+class LogoutView(View):
+    def get(self, request):
+        logout(request)
+        return redirect("login")
+
+
+class SignupView(View):
+    def get(self, request):
+        form = UserCreationForm()
+        return render(request, "signup.html", {"form": form})
+
+    def post(self, request):
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save()
+            login(request, user)
+            return redirect("event_list")
+        return render(request, "signup.html", {"form": form})
