+++ eventplannerdb/views/events/event_views.py
@@ -1,11 +1,8 @@
-from django.shortcuts import render, get_object_or_404, redirect
-from django.views import View
+from django.shortcuts import get_object_or_404, redirect
 from django.views.generic import ListView, DetailView, CreateView
-from django.contrib.auth import login, logout
 from django.contrib.auth.decorators import login_required
-from django.contrib.auth.forms import UserCreationForm, AuthenticationForm
-from .models import Event, Survey, Speaker, EventAgenda
-from .forms import SurveyForm
+from ...models import Event, Survey, Speaker, EventAgenda
+from ...forms import SurveyForm
 
 class EventListView(ListView):
     """
@@ -103,35 +100,3 @@ class EventAgendaCreateView(CreateView):
     success_url = "/events/"  # Redirect to the event list after agenda creation
 
 
-class LoginView(View):
-    def get(self, request):
-        form = AuthenticationForm()
-        return render(request, "login.html", {"form": form})
-
-    def post(self, request):
-        form = AuthenticationForm(data=request.POST)
-        if form.is_valid():
-            user = form.get_user()
-            login(request, user)
-            return redirect("event_list")
-        return render(request, "login.html", {"form": form})
-
-
-class LogoutView(View):
-    def get(self, request):
-        logout(request)
-        return redirect("login")
-
-
-class SignupView(View):
-    def get(self, request):
-        form = UserCreationForm()
-        return render(request, "signup.html", {"form": form})
-
-    def post(self, request):
-        form = UserCreationForm(request.POST)
-        if form.is_valid():
-            user = form.save()
-            login(request, user)
-            return redirect("event_list")
-        return render(request, "signup.html", {"form": form})
